requirements.txt
asgiref==3.9.2
Django==3.2.25
django-extensions==4.1
django-robots==6.1
django-taggit==3.1.0
django_debug_toolbar==3.8.1
legacy-cgi==2.6.3
pillow==12.0.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.3
